backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "funasr>=1.3.1",
    "openai-whisper>=20250625",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
